+++ projects/hipblaslt/tensilelite/Tensile/Tests/unit/test_helper_cache.py
@@ -57,7 +57,7 @@ def _write_test_files(tmp_path, cpp_content="void f(){}", h_content="#pragma onc
 
 class TestComputeCacheKey:
     def test_deterministic(self, tmp_path):
-        from Tensile.Toolchain.Source import _computeCacheKey
+        from Tensile.Toolchain.HelperKernelCache import _computeCacheKey
         kernel_path = _write_test_files(tmp_path)
         compiler = MockCompiler()
         k1 = _computeCacheKey(kernel_path, tmp_path, ["gfx942"], compiler)
@@ -66,7 +66,7 @@ class TestComputeCacheKey:
         assert len(k1) == 64  # sha256 hex digest
 
     def test_different_source_different_key(self, tmp_path):
-        from Tensile.Toolchain.Source import _computeCacheKey
+        from Tensile.Toolchain.HelperKernelCache import _computeCacheKey
         kernel_path = _write_test_files(tmp_path, cpp_content="void f(){}")
         compiler = MockCompiler()
         k1 = _computeCacheKey(kernel_path, tmp_path, ["gfx942"], compiler)
@@ -75,7 +75,7 @@ class TestComputeCacheKey:
         assert k1 != k2
 
     def test_different_arch_different_key(self, tmp_path):
-        from Tensile.Toolchain.Source import _computeCacheKey
+        from Tensile.Toolchain.HelperKernelCache import _computeCacheKey
         kernel_path = _write_test_files(tmp_path)
         compiler = MockCompiler()
         k1 = _computeCacheKey(kernel_path, tmp_path, ["gfx942"], compiler)
@@ -83,7 +83,7 @@ class TestComputeCacheKey:
         assert k1 != k2
 
     def test_arch_order_irrelevant(self, tmp_path):
-        from Tensile.Toolchain.Source import _computeCacheKey
+        from Tensile.Toolchain.HelperKernelCache import _computeCacheKey
         kernel_path = _write_test_files(tmp_path)
         compiler = MockCompiler()
         k1 = _computeCacheKey(kernel_path, tmp_path, ["gfx942", "gfx1100"], compiler)
@@ -91,28 +91,28 @@ class TestComputeCacheKey:
         assert k1 == k2
 
     def test_different_compiler_version_different_key(self, tmp_path):
-        from Tensile.Toolchain.Source import _computeCacheKey
+        from Tensile.Toolchain.HelperKernelCache import _computeCacheKey
         kernel_path = _write_test_files(tmp_path)
         k1 = _computeCacheKey(kernel_path, tmp_path, ["gfx942"], MockCompiler(version=(6, 0, 0)))
         k2 = _computeCacheKey(kernel_path, tmp_path, ["gfx942"], MockCompiler(version=(6, 1, 0)))
         assert k1 != k2
 
     def test_different_rocm_version_different_key(self, tmp_path):
-        from Tensile.Toolchain.Source import _computeCacheKey
+        from Tensile.Toolchain.HelperKernelCache import _computeCacheKey
         kernel_path = _write_test_files(tmp_path)
         k1 = _computeCacheKey(kernel_path, tmp_path, ["gfx942"], MockCompiler(rocm_version=(6, 0, 0)))
         k2 = _computeCacheKey(kernel_path, tmp_path, ["gfx942"], MockCompiler(rocm_version=(6, 1, 0)))
         assert k1 != k2
 
     def test_asan_changes_key(self, tmp_path):
-        from Tensile.Toolchain.Source import _computeCacheKey
+        from Tensile.Toolchain.HelperKernelCache import _computeCacheKey
         kernel_path = _write_test_files(tmp_path)
         k1 = _computeCacheKey(kernel_path, tmp_path, ["gfx942"], MockCompiler(asan=False))
         k2 = _computeCacheKey(kernel_path, tmp_path, ["gfx942"], MockCompiler(asan=True))
         assert k1 != k2
 
     def test_static_header_change_changes_key(self, tmp_path):
-        from Tensile.Toolchain.Source import _computeCacheKey
+        from Tensile.Toolchain.HelperKernelCache import _computeCacheKey
         kernel_path = _write_test_files(tmp_path)
         compiler = MockCompiler()
         k1 = _computeCacheKey(kernel_path, tmp_path, ["gfx942"], compiler)
@@ -123,23 +123,23 @@ class TestComputeCacheKey:
 
 class TestCheckCache:
     def test_returns_none_when_dir_missing(self, tmp_path):
-        from Tensile.Toolchain.Source import _checkCache
+        from Tensile.Toolchain.HelperKernelCache import _checkCache
         assert _checkCache(tmp_path, "nonexistent_hash") is None
 
     def test_returns_none_when_dir_empty(self, tmp_path):
-        from Tensile.Toolchain.Source import _checkCache
+        from Tensile.Toolchain.HelperKernelCache import _checkCache
         (tmp_path / "some_hash").mkdir()
         assert _checkCache(tmp_path, "some_hash") is None
 
     def test_returns_none_when_file_zero_size(self, tmp_path):
-        from Tensile.Toolchain.Source import _checkCache
+        from Tensile.Toolchain.HelperKernelCache import _checkCache
         entry = tmp_path / "some_hash"
         entry.mkdir()
         (entry / "Kernels.so-000-gfx942.hsaco").write_bytes(b"")
         assert _checkCache(tmp_path, "some_hash") is None
 
     def test_returns_files_on_valid_entry(self, tmp_path):
-        from Tensile.Toolchain.Source import _checkCache
+        from Tensile.Toolchain.HelperKernelCache import _checkCache
         entry = tmp_path / "some_hash"
         entry.mkdir()
         (entry / "Kernels.so-000-gfx942.hsaco").write_bytes(b"\x7fELF")
@@ -150,7 +150,7 @@ class TestCheckCache:
         assert all(f.suffix == ".hsaco" for f in result)
 
     def test_ignores_non_hsaco_files(self, tmp_path):
-        from Tensile.Toolchain.Source import _checkCache
+        from Tensile.Toolchain.HelperKernelCache import _checkCache
         entry = tmp_path / "some_hash"
         entry.mkdir()
         (entry / "Kernels.so-000-gfx942.hsaco").write_bytes(b"\x7fELF")
@@ -161,7 +161,7 @@ class TestCheckCache:
 
 class TestPopulateCache:
     def test_populates_empty_cache(self, tmp_path):
-        from Tensile.Toolchain.Source import _populateCache
+        from Tensile.Toolchain.HelperKernelCache import _populateCache
         cache_dir = tmp_path / "cache"
         cache_dir.mkdir()
         src_dir = tmp_path / "src"
@@ -174,7 +174,7 @@ class TestPopulateCache:
         assert cached.read_bytes() == b"\x7fELF_data_1"
 
     def test_skips_when_entry_exists(self, tmp_path):
-        from Tensile.Toolchain.Source import _populateCache
+        from Tensile.Toolchain.HelperKernelCache import _populateCache
         cache_dir = tmp_path / "cache"
         entry = cache_dir / "abc123"
         entry.mkdir(parents=True)
@@ -185,7 +185,7 @@ class TestPopulateCache:
         assert (entry / "Kernels.so-000-gfx942.hsaco").read_bytes() == b"original"
 
     def test_cleans_up_tmp_on_race(self, tmp_path):
-        from Tensile.Toolchain.Source import _populateCache
+        from Tensile.Toolchain.HelperKernelCache import _populateCache
         cache_dir = tmp_path / "cache"
         cache_dir.mkdir()
         # Pre-create the final dir to simulate a race
@@ -199,7 +199,7 @@ class TestPopulateCache:
         assert len(tmp_dirs) == 0
 
     def test_creates_cache_dir_if_missing(self, tmp_path):
-        from Tensile.Toolchain.Source import _populateCache
+        from Tensile.Toolchain.HelperKernelCache import _populateCache
         cache_dir = tmp_path / "cache" / "subdir"
         src = tmp_path / "f.hsaco"
         src.write_bytes(b"\x7fELF")
@@ -212,7 +212,7 @@ class TestBuildSourceCodeObjectFilesCache:
 
     def test_cache_miss_creates_entry(self, tmp_path, monkeypatch):
         """On cache miss, after compilation, cache dir should be populated."""
-        from Tensile.Toolchain.Source import _computeCacheKey, _checkCache
+        from Tensile.Toolchain.HelperKernelCache import _computeCacheKey, _checkCache
         cache_dir = tmp_path / "cache"
         monkeypatch.setenv("TENSILE_HELPER_CACHE_DIR", str(cache_dir))
         monkeypatch.delenv("TENSILE_DISABLE_HELPER_CACHE", raising=False)
@@ -238,7 +238,7 @@ class TestBuildSourceCodeObjectFilesCache:
 
     def test_cache_hit_copies_files(self, tmp_path):
         """Pre-populate cache, verify _checkCache finds it."""
-        from Tensile.Toolchain.Source import _checkCache
+        from Tensile.Toolchain.HelperKernelCache import _checkCache
         cache_dir = tmp_path / "cache"
         entry = cache_dir / "test_key"
         entry.mkdir(parents=True)
